ECG 5000 Dataset

The data was pre-processed in two steps: (1) Extract each heartbeat. (2) Make each heartbeat equal length using interpolation. This dataset was originally used in the paper "A general framework for never-ending learning from time series streams", DAMI 29(6). After that, 5,000 heartbeats were randomly selected. The patient has severe congestive heart failure and the class values were obtained by automated annotation.

Usage

data(ECG5000)

Format

A list containing the following components:

df

data.frame with the following variables:

class

Corresponding class level of “ECG” curves with 4 classes.

sample

Factor variable. In the TSC database, the first 500 values (sample="train") are used as the training sample and the remaining 4500 (sample="test") for testing.

x

fdata class object with n = 5000 curves (per row) in 140 discretization points (per column).

Source

https://physionet.org/cgi-bin/atm/ATM and http://timeseriesclassification.com/description.php?Dataset=ECG5000

Details

The original dataset for "ECG5000" is a 20-hour long ECG downloaded from Physionet. The name is BIDMC Congestive Heart Failure Database (chfdb) and it is record "chf07". It was originally published in "Goldberger AL, Amaral LAN, Glass L, Hausdorff JM, Ivanov PCh, Mark RG, Mietus JE, Moody GB, Peng C-K, Stanley HE. PhysioBank, PhysioToolkit, and PhysioNet: Components of a New Research Resource for Complex Physiologic Signals. Circulation 101(23)".